refactor: matchStrings
